refactor: restructure CreateKiwi components by separating modal and card functionality into distinct files
