+++ apps/web/src/routes/my-kiwis/-kiwis/create-kiwi/index.ts
@@ -1,0 +1,2 @@
+export * from "./components";
+export { default as CreateKiwiModal } from "./modal";
